cleaned up and connected shuffle component to answerlist
