+++ src/App.js
@@ -1,11 +1,9 @@
 import React, { Component } from "react";
 import "./App.css";
 import store from "./store";
-import ShuffleContainer from "./ShuffleContainer";
-
 import DogsBreedImagesContainer from "./components/DogsBreedImagesContainer";
 import { Provider } from "react-redux";
-import AnswerListContainer from "./AnswerListContainer";
+import AnswerListContainer from "./components/AnswerListContainer";
 
 class App extends Component {
   render() {
@@ -14,10 +12,6 @@ class App extends Component {
         <div className="App">
           <DogsBreedImagesContainer />
           <h1>Breaking Breeds</h1>
-          <div>option 1</div>
-          <div>option 2</div>
-          <div>option 3</div>
-          <ShuffleContainer />
           <AnswerListContainer />
         </div>
       </Provider>
