--- src/components/Awnsers.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-const Awnsers = () => {
-  return (
-    <ul>
-      <li>Awnsers 1</li>
-      <li>Awnsers 2</li>
-      <li>Awnsers 3</li>
-    </ul>
-  );
-};
-
-export default Awnsers;
