--- src/components/Image.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-const Image = () => {
-  return (
-    <img
-      src="https://static.valentinosdisplays.com/originalimages/medium/2741-dog-mannequin-uk.jpg"
-      alt="dummy dog"
-    />
-  );
-};
-
-export default Image;
